requirements.txt
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.4
cachetools==5.3.1
cfgv==3.4.0
click==8.1.7
distlib==0.3.7
fastapi==0.101.1
filelock==3.12.2
h11==0.14.0
identify==2.5.27
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
nodeenv==1.8.0
platformdirs==3.10.0
pre-commit==3.3.3
psycopg2-binary==2.9.7
pydantic==2.2.1
pydantic-settings==2.0.3
pydantic_core==2.6.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
tzlocal==5.0.1
uvicorn==0.23.2
virtualenv==20.24.3
